+++ src/model/label.go
@@ -87,12 +87,14 @@ func (m *Label) Find(ctx context.Context, productID int64, pg tpl.Pagination) ([
 	labels := make([]schema.Label, 0)
 	cursor := pg.TokenToID()
 	db := m.DB.Where("`id` <= ? and `product_id` = ? and `offline_at` is null", cursor, productID)
+	dbc := m.DB.Where("`product_id` = ? and `offline_at` is null", productID)
 	if pg.Q != "" {
 		db = m.DB.Where("`id` <= ? and `product_id` = ? and `offline_at` is null and `name` like ?", cursor, productID, pg.Q)
+		dbc = m.DB.Where("`product_id` = ? and `offline_at` is null and `name` like ?", productID, pg.Q)
 	}
 
 	total := 0
-	err := db.Model(&schema.Label{}).Count(&total).Error
+	err := dbc.Model(&schema.Label{}).Count(&total).Error
 	if err == nil {
 		err = db.Order("`id` desc").Limit(pg.PageSize + 1).Find(&labels).Error
 	}